types: move types to separate folder
