+++ src/looking_glass/cheshire-cat.ts
@@ -3,9 +3,10 @@ import type { Embeddings } from '@langchain/core/embeddings'
 import type { BaseLanguageModel } from '@langchain/core/language_models/base'
 import { getEmbedder, getLLM } from '@factory'
 import { type Form, type Tool, isForm, isTool, madHatter } from '@mh'
-import { type PointData, type VectorMemory, getVectorMemory } from '@memory'
+import { type VectorMemory, getVectorMemory } from '@memory'
 import { db } from '@db'
 import { log } from '@logger'
+import type { PointData } from '@dto/vector-memory.ts'
 import { AgentManager } from './agent-manager.ts'
 import { StrayCat } from './stray-cat.ts'
